move lib_bilagrid and remove custom betas
